Move `block_arrival` class to its own file
